+++ Product-Service/src/main/java/ru/svanchukov/productservice/service/ProductsService.java
@@ -1,6 +1,5 @@
 package ru.svanchukov.productservice.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,6 +8,8 @@ import ru.svanchukov.productservice.dto.product.CreateNewProductDTO;
 import ru.svanchukov.productservice.dto.product.ProductDTO;
 import ru.svanchukov.productservice.dto.product.UpdateProductDTO;
 import ru.svanchukov.productservice.entity.Product;
+import ru.svanchukov.productservice.handler.ProductNotFoundException;
+import ru.svanchukov.productservice.handler.ProductSavingException;
 import ru.svanchukov.productservice.repository.ProductRepository;
 
 import java.util.List;
@@ -25,7 +26,6 @@ public class ProductsService {
     private static final Logger LOGGER = LoggerFactory.getLogger(ProductService.class);
 
     private final ProductRepository productRepository;
-    private final ObjectMapper objectMapper;
 
     /**
      * Создание и сохранение нового продукта.
@@ -45,7 +45,7 @@ public class ProductsService {
             LOGGER.info("Продукт с именем {} успешно сохранен", createNewProductDTO.getName());
         } catch (Exception e) {
             LOGGER.error("Ошибка при сохранении продукта: {}", createNewProductDTO.getName(), e);
-            throw new RuntimeException("Ошибка при сохранении продукта", e);
+            throw new ProductSavingException("Ошибка при сохранении продукта");
         }
 
         return mapToDto(product);
@@ -94,11 +94,11 @@ public class ProductsService {
      */
     public void delete(Long productId) {
         LOGGER.info("Запрос на удаление продукта с ID: {}", productId);
-        Product product = productRepository.findById(productId)
-                .orElseThrow(() -> {
-                    LOGGER.error("Продукт с ID {} не найден для удаления", productId);
-                    return new RuntimeException("Продукт с ID " + productId + " не найден");
-                });
+
+        if (!productRepository.existsById(productId)) {
+            LOGGER.error("Продукт с ID {} не найден для удаления", productId);
+            throw new ProductNotFoundException("Продукт с ID " + productId + " не найден");
+        }
 
         productRepository.deleteById(productId);
         LOGGER.info("Продукт с ID: {} успешно удален", productId);
